backend/requirements.txt
asgiref
attrs
Babel
beautifulsoup4
certifi
charset-normalizer
colorama
config
datedelta
Django
django-allauth
django-cors-headers
django-cors-middleware
django-debug-toolbar
django-js-asset
django-localflavor
django-mathfilters
django-mptt
django-no-last-login
django-simple-history
django-smtp-ssl
django-templatetags
django-utils
djangorestframework
gusregon
idna
imgkit
iniconfig
isodate
Jinja2
litex
litex.regon
lxml
Markdown
MarkupSafe
migen
numpy
packaging
pandas
pdfkit
pillow
platformdirs
pluggy
psycopg-binary
psycopg2-binary
pyserial
pytesseract
pytest
pytest-django
python-dateutil
python-stdnum
pytz
PyVirtualDisplay
requests
requests-file
requests-toolbelt
setuptools
six
soupsieve
sqlparse
tzdata
urllib3
zeep
